pyproject.toml
[project]
name = "konkurrenttracker"
version = "0.1.0"
description = "Konkurrenceovervaagning for Epico - ugentlig PDF-rapport + live dashboard"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "feedparser>=6.0.11",
    "anthropic>=0.39.0",
    "jinja2>=3.1.0",
    "structlog>=24.4.0",
    "pyyaml>=6.0.2",
    "apscheduler>=3.10.4",
    "tzdata>=2024.1",
    "weasyprint>=63.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"
